requirements.txt
Flask-RESTful==0.3.6
Flask-SQLAlchemy>=2.3.2
pymongo>=3.6.1
aiomysql>=0.0.14
PyMySQL>=0.8.1
Scrapy>=1.5.0
pytest>=3.5.1
pytest-twisted>=1.8
Jinja2>=2.10
regex>=2018.2.21
sqlalchemy>=1.2.7
pymysql
sqlalchemy_utils